refactor: ajustes e melhorias na estrutura do ServiceManager
