+++ Api-Network/PrimeiroAppNetwork/PrimeiroAppNetwork/Service/HTTPMethod.swift
@@ -1,0 +1,15 @@
+//
+//  HTTPMethod.swift
+//  PrimeiroAppNetwork
+//
+//  Created by Jenifer Rocha on 22/05/25.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case delete = "DELETE"
+}
